pyproject.toml
[project]
name = "langchain-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.17",
    "langchain-core>=0.1.50",
    "langchain-google-genai>=0.0.6",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "regex>=2024.11.6",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
]
